Update scalar advdiff unit test to use a skewed hex element

+++ unit_tests/UnitTestSuppAlgDataSharing.C
@@ -127,7 +127,7 @@ public:
           sierra::nalu::ScratchViews prereqData(team, bulkData_, topo, dataNeededBySuppAlgs_);
 
           // See get_num_bytes_pre_req_data for padding
-          EXPECT_EQ(bytes_per_thread, prereqData.total_bytes() + 8 * sizeof(double));
+          EXPECT_EQ(static_cast<unsigned>(bytes_per_thread), prereqData.total_bytes() + 8 * sizeof(double));
 
           Kokkos::parallel_for(Kokkos::TeamThreadRange(team, bkt.size()), [&](const size_t& jj)
           {
